Check AMS-II.E fuel switching criterion

Projects that replace fossil fuels with lower-carbon options must demonstrate that the weighted emission factor of the project fuel mix is below the baseline mix.

check_applicability_fuel_switching(
  baseline_data,
  project_data,
  energy_col = "baseline_thermal_energy_gj",
  emission_factor_col = "baseline_thermal_emission_factor_tco2_per_gj",
  project_energy_col = "project_thermal_energy_gj",
  project_emission_factor_col = "project_thermal_emission_factor_tco2_per_gj"
)

Arguments

baseline_data

Tibble containing baseline thermal energy and emission factors.

project_data

Tibble containing project thermal energy and emission factors.

energy_col

Baseline column storing thermal energy demand (GJ).

emission_factor_col

Baseline column storing emission factors (tCO2e/GJ).

project_energy_col

Project column storing thermal energy demand (GJ).

project_emission_factor_col

Project column storing emission factors (tCO2e/GJ).

Value

Logical scalar indicating whether the project fuel mix is lower in emissions than the baseline mix.

Examples

baseline <- tibble::tibble(thermal_energy_gj = c(320, 180), emission_factor_tco2_per_gj = c(0.07, 0.072))
project <- tibble::tibble(thermal_energy_gj = c(280, 120), emission_factor_tco2_per_gj = c(0.045, 0.05))
check_applicability_fuel_switching(baseline, project,
  energy_col = "thermal_energy_gj",
  emission_factor_col = "emission_factor_tco2_per_gj",
  project_energy_col = "thermal_energy_gj",
  project_emission_factor_col = "emission_factor_tco2_per_gj"
)
#> [1] TRUE